requirements.txt
aiogram==3.5.0
aiohttp>=3.9.0,<3.10.0
fastapi>=0.110.0
uvicorn==0.21.0
python-dotenv==1.0.1
sqlalchemy>=2.0.0
aiosqlite==0.19.0
alembic>=1.11.0,<1.12.0